Refactor NotebookExplorer => SearchManager
